feat(post): 게시물 작성자 권한 검사 추가 - 본인 게시물만 수정·삭제 가능(실습)

+++ src/main/java/kr/co/bullets/board/service/PostService.java
@@ -1,6 +1,8 @@
 package kr.co.bullets.board.service;
 
 import kr.co.bullets.board.exception.post.PostNotFoundException;
+import kr.co.bullets.board.exception.user.UserNotAllowedException;
+import kr.co.bullets.board.model.entity.UserEntity;
 import kr.co.bullets.board.model.post.Post;
 import kr.co.bullets.board.model.post.PostPatchRequestBody;
 import kr.co.bullets.board.model.post.PostPostRequestBody;
@@ -23,37 +25,38 @@ public class PostService {
 
   public Post getPostByPostId(Long postId) {
     var postEntity =
-        postEntityRepository
-            .findById(postId)
-            .orElseThrow(
-                () -> new PostNotFoundException(postId));
+        postEntityRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
     return Post.from(postEntity);
   }
 
-  public Post createPost(PostPostRequestBody postPostRequestBody) {
-    var postEntity = new PostEntity();
-    postEntity.setBody(postPostRequestBody.body());
-    var savedPostEntity = postEntityRepository.save(postEntity);
+  public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+    var savedPostEntity =
+        postEntityRepository.save(PostEntity.of(postPostRequestBody.body(), currentUser));
     return Post.from(savedPostEntity);
   }
 
-  public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+  public Post updatePost(
+      Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
     var postEntity =
-        postEntityRepository
-            .findById(postId)
-            .orElseThrow(
-                () -> new PostNotFoundException(postId));
+        postEntityRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+    if (!postEntity.getUser().equals(currentUser)) {
+      throw new UserNotAllowedException();
+    }
+
     postEntity.setBody(postPatchRequestBody.body());
-    var updatedPostEntity = postEntityRepository.save(postEntity);
-    return Post.from(updatedPostEntity);
+    var updatedEntity = postEntityRepository.save(postEntity);
+    return Post.from(updatedEntity);
   }
 
-  public void deletePost(Long postId) {
+  public void deletePost(Long postId, UserEntity currentUser) {
     var postEntity =
-        postEntityRepository
-            .findById(postId)
-            .orElseThrow(
-                () -> new PostNotFoundException(postId));
+        postEntityRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+    if (!postEntity.getUser().equals(currentUser)) {
+      throw new UserNotAllowedException();
+    }
+
     postEntityRepository.delete(postEntity);
   }
 }
